+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/TemporalFieldPanel.java
@@ -89,7 +89,7 @@ public final class TemporalFieldPanel<T extends Temporal> extends JPanel {
   /**
    * @param transferFocusOnEnter specifies whether focus should be transferred on Enter
    */
-  public void setTransferFocusOnEnter(boolean transferFocusOnEnter) {
+  public void transferFocusOnEnter(boolean transferFocusOnEnter) {
     if (transferFocusOnEnter) {
       TransferFocusOnEnter.enable(temporalField);
       if (calendarButton != null) {
@@ -259,7 +259,7 @@ public final class TemporalFieldPanel<T extends Temporal> extends JPanel {
 
     @Override
     protected void enableTransferFocusOnEnter(TemporalFieldPanel<T> component) {
-      component.setTransferFocusOnEnter(true);
+      component.transferFocusOnEnter(true);
     }
 
     @Override