+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/TextFieldPanel.java
@@ -127,7 +127,7 @@ public final class TextFieldPanel extends JPanel {
   /**
    * @param transferFocusOnEnter specifies whether focus should be transferred on Enter
    */
-  public void setTransferFocusOnEnter(boolean transferFocusOnEnter) {
+  public void transferFocusOnEnter(boolean transferFocusOnEnter) {
     if (transferFocusOnEnter) {
       TransferFocusOnEnter.enable(textField);
       TransferFocusOnEnter.enable(button);
@@ -391,7 +391,7 @@ public final class TextFieldPanel extends JPanel {
 
     @Override
     protected void enableTransferFocusOnEnter(TextFieldPanel component) {
-      component.setTransferFocusOnEnter(true);
+      component.transferFocusOnEnter(true);
     }
   }