+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/icon/DefaultFrameworkIcons.java
@@ -190,8 +190,8 @@ public final class DefaultFrameworkIcons implements FrameworkIcons, Logos {
   }
 
   @Override
-  public void setIconColor(Color color) {
-    icons.setIconColor(color);
+  public void iconColor(Color color) {
+    icons.iconColor(color);
   }
 
   @Override